New plugin: traverse
This adds a plugin for traversing across windows based on their position
in global space. By "plugin" this means an importable module that
exports functions that can be used as-is (via `lazy.function`) rather
than a full config. I've added a table for this to the readme to set the
precedent.